fix: change evaluation method to lazy evaluation
