+++ src/index.ts
@@ -34,7 +34,6 @@ import pickBy from "./pickBy";
 import pipe from "./pipe";
 import pipe1 from "./pipe1";
 import reduce from "./reduce";
-import reverse from "./reverse";
 import size from "./size";
 import some from "./some";
 import sort from "./sort";
@@ -81,7 +80,6 @@ export {
   pipe,
   pipe1,
   reduce,
-  reverse,
   size,
   some,
   sort,
