+++ test/Lazy/reverse.spec.ts
@@ -1,4 +1,6 @@
-import { pipe, toAsync, reverse } from "../src";
+import { pipe, toAsync, reverse, toArray, filter, map, delay } from "../../src";
+import concurrent, { Concurrent } from "../../src/Lazy/concurrent";
+import { callFuncAfterTime, generatorMock } from "../utils";
 
 describe("reverse", function () {
   describe("sync", function () {
@@ -7,16 +9,16 @@ describe("reverse", function () {
         [1, 2, 3, 4],
         [4, 3, 2, 1],
       ],
-      ["abcd", "dcba"],
+      ["abcd", ["d", "c", "b", "a"]],
     ])(
       "should return the given elements %s in reverse order %s",
       function (input, result) {
-        expect(reverse(input)).toEqual(result);
+        expect(toArray(reverse(input))).toEqual(result);
       },
     );
 
     it("should be able to be used as a curried function in the pipeline", function () {
-      const res = pipe([1, 2, 3, 4], reverse);
+      const res = pipe([1, 2, 3, 4], reverse, toArray);
       expect(res).toEqual([4, 3, 2, 1]);
     });
   });
@@ -31,12 +33,12 @@ describe("reverse", function () {
     ])(
       "should return the given elements %s in reverse order %s",
       async function (input, result) {
-        expect(await reverse(toAsync(input as any))).toEqual(result);
+        expect(await toArray(reverse(toAsync(input as any)))).toEqual(result);
       },
     );
 
     it("should be able to be used as a curried function in the pipeline", async function () {
-      const res = await pipe([1, 2, 3, 4], toAsync, reverse);
+      const res = await pipe([1, 2, 3, 4], toAsync, reverse, toArray);
       expect(res).toEqual([4, 3, 2, 1]);
     });
   });
